+++ src/api/endpoints/__init__.py
@@ -3,3 +3,5 @@ from api.endpoints.user_endpoint import user_router
 from api.endpoints.group_endpoint import group_router
 from api.endpoints.crontab_endpoint import crontab_router
 from api.endpoints.storage_endpoint import storage_router
+from api.endpoints.host_endpoint import host_router
+from api.endpoints.hardware import hardware_router