+++ src/main.py
@@ -1,7 +1,7 @@
 """Main module for the API REST application."""
 from fastapi import FastAPI
 
-from api import user_router, group_router, crontab_router, storage_router
+from api import user_router, group_router, crontab_router, storage_router, host_router, hardware_router
 
 app = FastAPI()
 
@@ -9,3 +9,5 @@ app.include_router(user_router)
 app.include_router(group_router)
 app.include_router(crontab_router)
 app.include_router(storage_router)
+app.include_router(host_router)
+app.include_router(hardware_router)